move interpreter code out of main module __init__.py
