+++ tox.ini
@@ -16,7 +16,7 @@ use_parentheses=True
 
 [flake8]
 max-line-length= 100
-exclude= venv*,.tox,docs,build,wasm/__init__.py
+exclude= venv*,.tox,docs,build,wasm/main.py
 ignore=
 
 [testenv]
